# openapi.yaml
...
openapi: 3.1.0
info:
  title: aperIA - Python API
  description: '

    API do **aperIA** — ASPM (Application Security Posture Management) open-source.


    Um webhook de Pull Request do GitHub dispara um pipeline de scanners de

    segurança em 3 tiers (via Celery). O Claude raciocina como atacante

    (correlaciona findings, monta attack paths) e entrega patches como

    **GitHub code suggestions** para aprovação humana.


    > **Premissa inviolável:** o aperIA **nunca** aplica código sozinho.

    > Todo patch sai como code suggestion para aprovação humana.

    '
  contact:
    name: aperIA
  license:
    name: Open Source
  version: 1.0.0
paths:
...
  /users:
    post:
      tags:
      - users
      summary: Criar usuário
      description: 'Cria um novo usuário.


        A senha (mínimo 8 caracteres) é armazenada com hash Argon2 e o e-mail

        deve ser único. Retorna apenas o `id` do usuário recém-criado.'
      operationId: create_user_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Usuário criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreatedResponse'
        '400':
          description: Erro de validação de domínio.
          content:
            application/json:
              example:
                detail: E-mail invalido
        '409':
          description: E-mail já cadastrado.
          content:
            application/json:
              example:
                detail: E-mail já cadastrado
        '500':
          description: 'Erro interno (ex.: falha de permissão no banco).'
          content:
            application/json:
              example:
                detail: Internal server error
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /users/{user_id}:
    get:
      tags:
      - users
      summary: Consultar usuário por ID
      description: Retorna os dados públicos de um usuário a partir do seu UUID.
      operationId: get_user_users__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Usuário encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Usuário não encontrado.
          content:
            application/json:
              example:
                detail: Usuário não encontrado
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /repositories/{repository_id}/scan:
    post:
      tags:
      - repositories
      summary: Iniciar scan manual do repositório
      description: "Dispara, sob demanda, o **mesmo** pipeline que o webhook de PR dispara.\n\nO commit\
        \ analisado é o HEAD do `default_branch` do repositório, resolvido\nao vivo no GitHub com o installation\
        \ token. Como não há PR, o pipeline roda\ncom `pr_number=None`: os relatórios ficam na projeção\
        \ da API em vez de\nvirarem comentário de PR.\n\nContrato de erros:\n\n- `404` se o repositório\
        \ não existir **ou** não pertencer ao usuário (nunca\n  403 — não vazamos a existência do recurso).\n\
        - `409` se o repositório estiver `active=false`. Desativado significa \"não\n  analisar\": um\
        \ scan manual aqui seria contraditório, e responder 404\n  esconderia do usuário um repositório\
        \ que ele mesmo pode reativar via\n  `PATCH /repositories/{id}`. Por isso 409 com mensagem explícita.\n\
        - `409` se já existir um `ScanJob` em andamento (algum tier `queued`/\n  `running`) para o commit\
        \ resolvido — evita duplicar o pipeline.\n- `502` se o GitHub não responder o HEAD do branch.\n\
        - `503` se o GitHub App não estiver configurado (mesmo espírito do\n  `GET /github/connect`, mas\
        \ aqui o que falta são as credenciais do App —\n  `GITHUB_APP_ID` / `GITHUB_PRIVATE_KEY_PATH`\
        \ — necessárias para emitir o\n  installation token)."
      operationId: trigger_repository_scan_repositories__repository_id__scan_post
      security:
      - HTTPBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Repository Id
      responses:
        '202':
          description: Pipeline disparado para o HEAD do branch default.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualScanResponse'
              example:
                status: queued
                commit_sha: 9b2e1f0a
                branch: main
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Repositorio nao encontrado.
        '409':
          description: Repositorio desativado, ou scan já em andamento para o commit.
          content:
            application/json:
              examples:
                desativado:
                  summary: Repositório desativado
                  value:
                    detail: 'Repositorio desativado: reative antes de iniciar um scan.'
                em_andamento:
                  summary: Scan já rodando para o mesmo commit
                  value:
                    detail: Ja existe um scan em andamento para o commit 9b2e1f0a.
        '502':
          description: Falha ao consultar o GitHub para resolver o HEAD do branch.
        '503':
          description: GitHub App nao configurado.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /webhook/github:
    post:
      tags:
      - webhook
      summary: Webhook de eventos do GitHub
      description: 'Recebe eventos do GitHub (GitHub App) e dispara o pipeline de segurança.


        A assinatura HMAC-SHA256 do corpo é validada contra o header

        `X-Hub-Signature-256` (segredo `GITHUB_WEBHOOK_SECRET`); assinatura

        ausente/inválida retorna `401`.


        Apenas eventos `pull_request` com ação `opened` ou `synchronize` disparam o

        pipeline (`queued`). Payloads sem `installation.id` ou outros eventos são

        ignorados (`ignored`).'
      operationId: github_webhook_webhook_github_post
      responses:
        '200':
          description: Evento processado (`queued` dispara o pipeline; `ignored` caso contrário).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Github Webhook Webhook Github Post
              examples:
                queued:
                  summary: PR aberto/atualizado
                  value:
                    status: queued
                    commit_sha: 9b2e1f0a...
                ignored_event:
                  summary: Evento não relevante
                  value:
                    status: ignored
                    event: ping
                ignored_no_installation:
                  summary: Sem installation_id
                  value:
                    status: ignored
                    reason: no installation_id
        '401':
          description: Assinatura HMAC ausente ou inválida.
          content:
            application/json:
              example:
                detail: Invalid HMAC signature
      parameters:
      - name: X-Hub-Signature-256
        in: header
        required: true
        description: Assinatura HMAC-SHA256 do corpo bruto, no formato `sha256=<hex>`.
        schema:
          type: string
        example: sha256=7d38cdd689735b008b3c702edd92eea23791c5f6
      - name: X-GitHub-Event
        in: header
        required: false
        description: 'Tipo do evento GitHub (ex.: `pull_request`, `ping`).'
        schema:
          type: string
        example: pull_request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload de webhook do GitHub (apenas campos consumidos).
              properties:
                action:
                  type: string
                  example: opened
                installation:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 12345678
                repository:
                  type: object
                  properties:
                    clone_url:
                      type: string
                      example: https://github.com/org/repo.git
                    full_name:
                      type: string
                      example: org/repo
                pull_request:
                  type: object
                  properties:
                    number:
                      type: integer
                      example: 42
                    head:
                      type: object
                      properties:
                        sha:
                          type: string
                    base:
                      type: object
                      properties:
                        sha:
                          type: string
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ManualScanResponse:
      properties:
        status:
          type: string
          title: Status
        commit_sha:
          type: string
          title: Commit Sha
        branch:
          type: string
          title: Branch
      type: object
      required:
      - status
      - commit_sha
      - branch
      title: ManualScanResponse
      description: 'Aceite de um scan manual (`POST /repositories/{id}/scan`).


        No espirito da resposta do webhook (`{"status": "queued", "commit_sha": ...}`),

        acrescentando o `branch` cujo HEAD foi resolvido.'
...
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 255
          minLength: 1
          title: Username
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
        email:
          type: string
          maxLength: 500
          minLength: 3
          title: Email
      additionalProperties: false
      type: object
      required:
      - username
      - password
      - email
      title: UserCreate
    UserCreatedResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
      - id
      title: UserCreatedResponse
    UserResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        username:
          type: string
          title: Username
        email:
          type: string
          title: Email
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - username
      - email
      - created_at
      - updated_at
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...